+++ .github/CODEOWNERS
@@ -1,7 +1,7 @@
 # Each line is a file pattern followed by one or more owners.
 # See https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-syntax
 /src/analysis @DataBiosphere/broad-interactive-analysis
-/src/data/datasets.js @DataBiosphere/data-explorer-eng
+/src/constants/datasets.js @DataBiosphere/data-explorer-eng
 /src/libs/brand*.js @DataBiosphere/terra-cobranding
 /src/libs/colors.js @DataBiosphere/terra-cobranding
 /src/libs/ajax/Apps.ts @DataBiosphere/broad-interactive-analysis
